fix(components): [notification] restore offset calc and extract utils
